update to sim_utils and diffraction_generator
Updating calculate_profile_data (c_p_d) to use the multiplicites of the planes in the calculation which was ignored before, resturcured using a new method in sim utils. The correct intensity is now calculated.
Still to do - update the ProfileSimulation to work for c_p_d, since the update has changed variables etc.