Faster implementation using ogrid
When profiling the performance of diffraction pattern simulation, most time was
spent in this function.

This implementation runs more than three times faster since it reduces the
calculation effort through broadcasting and re-uses coordinate and
distance calculations.

Instead of using diffpy, it performs the equivalent transformation to cartesian
and calculation of the norm directly to allow these optimizations.