Support CIF files from ICSD
A recent CIF file pulled from https://icsd.fiz-karlsruhe.de/
for collection code 163723 contained the atom type symbol "Au0+"
instead of just Au. This lead to empty diffraction patterns.

With this change only the first alphabetical
part of the atom type symbol is used as an element, and a warning is
emitted if unknown elements are encountered.

See https://www.iucr.org/__data/iucr/cifdic_html/1/cif_core.dic/Iatom_type_symbol.html
for the definition of the atom type symbol.

It can be debated if unrecognized atom type symbols should actually
be a hard error since the current
code will quietly zero the contribution of sites with unknown atom types, which may
lead to plausible but wrong diffraction patterns if recognized and
unrecognized atom type symbols are mixed in a CIF file.